requirements.txt
requests
pymodbus==1.5.1
python-dotenv
six